Organized the base code following the templates. Need to complete them with the code in the 'Claude' folder
