--- marc/orbitalSim.h
@@ -1,36 +1,0 @@
-﻿/**
- * @brief Orbital simulation
- * @author Marc S. Ressl
- *
- * @copyright Copyright (c) 2022-2023
- */
-
-#ifndef ORBITALSIM_H
-#define ORBITALSIM_H
-
-/**
- * @brief Orbital body definition
- */
-struct OrbitalBody
-{
-    // Fill in your code here...
-
-
-};
-
-/**
- * @brief Orbital simulation definition
- */
-struct OrbitalSim
-{
-    // Fill in your code here...
-
-
-};
-
-OrbitalSim *constructOrbitalSim(float timeStep);
-void destroyOrbitalSim(OrbitalSim *sim);
-
-void updateOrbitalSim(OrbitalSim *sim);
-
-#endif
